R Function to rank Hospitals by State on mortality rates
a function called RANKALL that takes two arguments: an outcome name
(outcome) and a hospital rank- ing (num). The function reads the
outcome-of-care-measures.csv file and returns a 2-column data frame
containing the hospital in each state that has the ranking specified in
num for a given 30-day mortality rate of a disease (outcome argument).